@@ -1,67 +1,71 @@
 {
-  "name": "notion-workspace",
-  "version": "18.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test",
-    "lint": "ng lint",
-    "ngx-notion-cms:lint": "ng lint ngx-notion-cms",
-    "ngx-notion-cms:test": "ng test ngx-notion-cms --no-watch --no-progress --browsers=ChromeHeadless",
-    "ngx-notion-cms:tailwind": "npx tailwind -c projects/ngx-notion-cms/tailwind.config.js -i projects/ngx-notion-cms/src/styles.css -o dist/ngx-notion-cms/styles.css",
-    "ngx-notion-cms:build": "ng build ngx-notion-cms",
-    "ngx-notion-cms:publish": "npm run ngx-notion-cms:lint && npm run ngx-notion-cms:build && npm run ngx-notion-cms:tailwind && npm publish dist/ngx-notion-cms/",
-    "serve:ssr:ngx-notion-testbed-app": "node dist/ngx-notion-testbed-app/server/server.mjs",
-    "build:prerender:ngx-notion-testbed-app": "npm run prerender && ng build ngx-notion-testbed-app",
-    "prerender": "node scripts/prerender-routes.js",
-    "prepare": "husky"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^18.2.0",
-    "@angular/common": "^18.2.0",
-    "@angular/compiler": "^18.2.0",
-    "@angular/core": "^18.2.0",
-    "@angular/forms": "^18.2.0",
-    "@angular/platform-browser": "^18.2.0",
-    "@angular/platform-browser-dynamic": "^18.2.0",
-    "@angular/platform-server": "^18.2.0",
-    "@angular/router": "^18.2.0",
-    "@angular/ssr": "^18.2.5",
-    "express": "^4.18.2",
-    "ngx-highlightjs": "^12.0.0",
-    "ngx-notion-cms": "^0.0.0-beta",
-    "rxjs": "~7.8.0",
-    "tslib": "^2.3.0",
-    "zone.js": "~0.14.10"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^18.2.5",
-    "@angular/cli": "^18.2.1",
-    "@angular/compiler-cli": "^18.2.0",
-    "@eslint/js": "^9.11.1",
-    "@types/express": "^4.17.17",
-    "@types/jasmine": "~5.1.0",
-    "@types/node": "^18.18.0",
-    "angular-eslint": "18.3.1",
-    "autoprefixer": "^10.4.20",
-    "eslint": "^9.11.1",
-    "globals": "^15.10.0",
-    "husky": "^8.0.0",
-    "jasmine-core": "~5.2.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.2.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.1.0",
-    "ng-packagr": "^18.2.0",
-    "postcss": "^8.4.47",
-    "prettier": "^3.3.3",
-    "prettier-eslint": "^16.3.0",
-    "tailwindcss": "^3.4.13",
-    "typescript": "~5.5.2",
-    "typescript-eslint": "^8.2.0"
-  }
+    "name": "notion-workspace",
+    "version": "18.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "start:hmr": "NG_HMR_TEMPLATES=1 ng serve",
+        "build": "ng build",
+        "build:prod": "ng build --configuration production",
+        "watch": "ng build --watch --configuration development",
+        "test": "ng test",
+        "lint": "ng lint",
+        "prepare": "husky",
+        "migrate:signals": "ng generate @angular/core:signals",
+        "ngx-notion-cms:lint": "ng lint ngx-notion-cms",
+        "ngx-notion-cms:test": "ng test ngx-notion-cms --no-watch --no-progress --browsers=ChromeHeadless",
+        "ngx-notion-cms:build": "ng build ngx-notion-cms --configuration production",
+        "ngx-notion-cms:tailwind": "npx tailwindcss -c projects/ngx-notion-cms/tailwind.config.js -i projects/ngx-notion-cms/src/styles.css -o dist/ngx-notion-cms/styles.css",
+        "ngx-notion-cms:publish": "npm run ngx-notion-cms:lint && npm run ngx-notion-cms:build && npm run ngx-notion-cms:tailwind && npm publish dist/ngx-notion-cms/",
+        "serve:ssr:ngx-notion-testbed-app": "node dist/ngx-notion-testbed-app/server/server.mjs",
+        "build:prerender:ngx-notion-testbed-app": "npm run prerender && ng build ngx-notion-testbed-app",
+        "prerender": "node scripts/prerender-routes.js",
+        "prepublishOnly": "npm run lint && npm run test"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^19.2.6",
+        "@angular/common": "^19.2.6",
+        "@angular/compiler": "^19.2.6",
+        "@angular/core": "^19.2.6",
+        "@angular/forms": "^19.2.6",
+        "@angular/platform-browser": "^19.2.6",
+        "@angular/platform-browser-dynamic": "^19.2.6",
+        "@angular/platform-server": "^19.2.6",
+        "@angular/router": "^19.2.6",
+        "@angular/ssr": "^19.2.7",
+        "express": "^4.18.2",
+        "ngx-highlightjs": "^12.0.0",
+        "ngx-notion-cms": "^0.0.0-beta",
+        "rxjs": "~7.8.0",
+        "tslib": "^2.3.0",
+        "zone.js": "~0.15.0"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^19.2.7",
+        "@angular/cli": "^19.2.7",
+        "@angular/compiler-cli": "^19.2.6",
+        "@eslint/js": "^9.11.1",
+        "@types/express": "^4.17.17",
+        "@types/jasmine": "~5.1.0",
+        "@types/node": "^18.18.0",
+        "angular-eslint": "18.3.1",
+        "autoprefixer": "^10.4.20",
+        "eslint": "^9.11.1",
+        "globals": "^15.10.0",
+        "husky": "^8.0.0",
+        "jasmine-core": "~5.2.0",
+        "karma": "~6.4.0",
+        "karma-chrome-launcher": "~3.2.0",
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.1.0",
+        "ng-packagr": "^19.2.2",
+        "postcss": "^8.4.47",
+        "prettier": "^3.3.3",
+        "prettier-eslint": "^16.3.0",
+        "tailwindcss": "^3.4.13",
+        "typescript": "~5.5.2",
+        "typescript-eslint": "^8.2.0"
+    }
 }
