@@ -7,9 +7,8 @@ import { GithubComponent } from "./components/github/github.component";
 
 @Component({
     selector: 'app-root',
-    standalone: true,
     imports: [NavComponent, RouterOutlet, ToggleButtonComponent, LogoComponent, GithubComponent],
     templateUrl: './app.component.html',
-    styleUrl: './app.component.css',
+    styleUrl: './app.component.css'
 })
 export class AppComponent { }
