@@ -4,11 +4,10 @@ import { NotionDatabaseComponent, NotionDatabaseItem } from 'ngx-notion-cms';
 
 @Component({
     selector: 'app-posts-view',
-    standalone: true,
     imports: [NotionDatabaseComponent, RouterModule],
     templateUrl: './posts-view.component.html',
     styleUrl: './posts-view.component.css',
-    changeDetection: ChangeDetectionStrategy.OnPush,
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class PostsViewComponent {
     private router = inject(Router);
